feat(profile): add memberManager
